requirements.txt
gitdb==4.0.5
GitPython==3.1.8
smmap==3.0.4
